Convert ZoneRecord Tier field to float, ns1 API now returns float values for this field

+++ rest/model/dns/zone.go
@@ -48,7 +48,7 @@ type ZoneRecord struct {
 	ID       string   `json:"id,omitempty"`
 	Link     string   `json:"link,omitempty"`
 	ShortAns []string `json:"short_answers,omitempty"`
-	Tier     int      `json:"tier,omitempty"`
+	Tier     float64  `json:"tier,omitempty"`
 	TTL      int      `json:"ttl,omitempty"`
 	Type     string   `json:"type,omitempty"`
 }
